requirements.txt
gradio
openai
SpeechRecognition
gTTS
python-dotenv
